2018/11/23
충돌판정에 필요없는 폴리곤의 번호가 아닌 방향을 인수로 받도록 변경
반사,기울기 가속 조건 변경
배열 정렬방식 변경
충돌 오차를 +-1도로 다시 복구

+++ Task_MapSide.h
@@ -56,7 +56,7 @@ namespace Map3d
 		//追加したい変数・メソッドはここに追加する	
 		//変数
 		Bbox arr[8][8];
-		int sizeZ, sizeX;
+		size_t sizeZ, sizeX;
 		int mapSize;
 		int sideNumber;
 		float chipSize;
